Client can page through millions of videos
Adress shortcoming of Kaltura to page through more than 10000 media
entries.

This is an initial draft, I noticed a difference between reported total
of media entries, and the number of entries being returned (it returned
more than the total, so there might be an interval overlap problem).

Further work is required to:
1) Make accurate
2) Clean up interface of the client (segregate)
3) Actually use it for something useful