Add updateDefaultThumbnailWithMiddleFrame
We want to be able to explicitly set up all videos to have a default
thumbnail image which is generated from slicing up the video and getting
the middle frame. Once it is done, we can switch to use the
`linkBuilder.getDefaultThumbnailUrl` when we create the playback
thumbnail urls which will enable us to be able to manually change the
thumbnails in the future.

[#171721039]