Ensure post field params are passed to the request
The updated unirest post field calls are now pure functions, i.e. they
do not have side effects on the original request object so the output
must be assigned so as not to lose it.

Also update the ObjectMapper to handle Strings as a special case to
match the previous behaviour.

Signed-off-by: Alex Glen <alex@boclips.com>